답변삭제에서 질문 삭제못하도록 수정

+++ src/main/java/com/codecrafter/commenting/service/ConversationService.java
@@ -218,6 +218,9 @@ public class ConversationService {
 	@Transactional
 	public void deleteAnswer(Long answerId) {
 		Conversation answer = findConversationById(answerId);
+		if(answer.isQuestion()) {
+			throw new IllegalArgumentException("질문은 삭제할 수 없습니다.");
+		}
 		conversationRepository.delete(answer);
 	}
